moves admin to package / interface
